requirements.txt
aiofiles==24.1.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
babel==2.17.0
bcrypt==4.3.0
black==25.1.0
certifi==2025.1.31
cfgv==3.4.0
click==8.1.8
databases==0.9.0
Deprecated==1.2.18
distlib==0.3.9
dnspython==2.7.0
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-cli==0.0.7
filelock==3.17.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
identify==2.6.8
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.5
limits==4.0.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
nodeenv==1.9.1
orjson==3.10.15
packaging==24.2
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
pre-commit==4.1.0
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-extra-types==2.10.2
pydantic-settings==2.8.0
pygments==2.19.1
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
tomli==2.2.1
typer==0.15.1
typing-extensions==4.12.2
ujson==5.10.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.2
watchfiles==1.0.4
websockets==15.0
wrapt==1.17.2
